# math: p_popcount_u64: Add function implementation