An adaptation of p_popcount_u32 methodology, with:
- masks of 64 bit length
- final shift factor changed from 24 (32-8) to 56 (64-8)

Resolves #33.

Signed-off-by: Mateusz Kacprzak <mateusz.kacprzak@yandex.ru>