@aolofsson 2015-03-07
Very nice! How fast is it? Have you tested it on x86?