@mateunho 2015-03-07
Complexity domain:
It's `12*n` operations for vector of length `n`, which for large vectors gives `O(n)` linear complexity - though, that's obvious.

Time domain: