@mateunho 2015-03-07
Complexity domain:
It's `12*n` operations for vector of length `n`, which, in theory, for large vectors gives `O(n)` linear complexity - though, that's obvious.

Time domain:
- tested current impl vs. naive (`var &= var - 1`)
- pseudo-random input
- 1k repetitions each vector size

vector size (n) | impl | avg. time (usec) | n / t | # faster (current/naive)
------: | :----- | ------: | ---: | ---:
1000 | current | 36 | 27.8 | 6
 | naive | 218 | 4.6 | -
10 000 | current | 422 | 23.7 | 5.5
 | naive | 2305 | 4.3 | -
100 000 | current | 6015 | 16.6 | 4.3
 | naive | 25 363 | 3.9 | -
1 000 000 | current | 19 690 | 50.8 | 5.1
 | naive | 100 678 | 9.9 | -

![image here](http://s8.postimg.org/5g9oiv6mt/curvsnai.png)
If those numbers were able to talk by themselves, they would say: "Pretty good".