@aolofsson 2015-03-07
Very nice!! Can you add a .dat file to the tests directory? The goal is to get the test suite also integrated with the autotools/travis.