@mateunho 2015-03-07
I'll add the test file, but test_main.c needs to be rewritten to read `uint64_t` and other non-float types from files.